pyproject.toml
[project]
name = "hse-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.127.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.45",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
